% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{elbow_plot_kmeans}
\alias{elbow_plot_kmeans}
\title{Spectral clustering}
\usage{
elbow_plot_kmeans(x, kmax = 15)
}
\arguments{
\item{x}{First value}
}
\value{
vector of groups
}
\description{
Spectral clustering
}
