% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_rr}
\alias{fit_rr}
\title{Fit a single ridge regression model}
\usage{
fit_rr(X, y, lambda)
}
\arguments{
\item{X}{First value}

\item{y}{Second value}

\item{lambda}{}
}
\value{
Vector of penalised regression coefficients
}
\description{
Fit a single ridge regression model
}
