% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_ocv_once}
\alias{get_ocv_once}
\title{Calculate leave one out cross validation error (OCV) for a single regression model}
\usage{
get_ocv_once(X, y, lambda)
}
\arguments{
\item{X}{Data matrix}

\item{y}{Column matrix of responses}

\item{lambda}{Numeric hyperparameter controlling the strength of the L2 penalisation (non-negative)}
}
\value{
Numeric OCV
}
\description{
Calculate leave one out cross validation error (OCV) for a single regression model
}
