% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optim_rr}
\alias{optim_rr}
\title{Find the optimal regularisation parameter through optimised leave one out cross validation}
\usage{
optim_rr(X, y, lams)
}
\arguments{
\item{v1}{First value}

\item{v2}{Second value}
}
\value{
Product of v1 and v2
}
\description{
Find the optimal regularisation parameter through optimised leave one out cross validation
}
