% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{par_reg}
\alias{par_reg}
\title{Fit a ridge regression model to multiple groups in parallel}
\usage{
par_reg(X, y, lams, idx)
}
\arguments{
\item{X}{Data matrix}

\item{y}{Column matrix of responses}

\item{lams}{Vector of regularisation parameters to test}

\item{idx}{Vector of sample groups}
}
\value{
List with two objects
lambdas A vector of the optimal value of lambda for each group
betas A matrix where columns are the fitted regression coefficients for each group
}
\description{
Fit a ridge regression model to multiple groups in parallel
}
