% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{pca}
\alias{pca}
\title{Our implementation of principal component analysis (PCA)}
\usage{
pca(x, sigma = 1.5)
}
\arguments{
\item{x}{numeric matrix of data, where rows are samples}

\item{sigma}{}
}
\value{
dataframe of sample projections onto PC1 and PC2
}
\description{
Given an data matrix x, a linear projection is applied to maximise sample variation.
The first two prinicpal components are returned.
}
