% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Spectral clustering}
\usage{
plot_clusters(data, clusters)
}
\arguments{
\item{x}{First value}
}
\value{
vector of groups
}
\description{
Spectral clustering
}
