% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_groups}
\alias{predict_groups}
\title{Predict new samples using the results from par_reg}
\usage{
predict_groups(X, betas, idx)
}
\arguments{
\item{X}{Data matrix of test samples}

\item{betas}{Matrix of regression coefficients}

\item{idx}{Vector of sample groups, corresponding to the columns of betas
(e.g. idx=c(1, 3) means betas[,1] will be used to predict X[1,], and betas[,3] will be used to predict X[2,])}
}
\value{
Vector of fitted values
}
\description{
Predict new samples using the results from par_reg
}
