% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_rr}
\alias{predict_rr}
\title{Predict new sample responses using a tuned regression model}
\usage{
predict_rr(X, beta)
}
\arguments{
\item{X}{Data matrix of test samples}

\item{beta}{Vector of regression coefficients}
}
\value{
Vector of fitted values
}
\description{
Predict new sample responses using a tuned regression model
}
