\name{RcppRidge-package}
\alias{RcppRidge}
\docType{package}
\title{
  Parallel Bayesian Ridge Regression
}
\description{
  A more detailed description of what the package does. A length
  of about one to five lines is recommended.
}
\details{
	This package was developed for a group project (Bristol Compass CDT), to perform Bayesian ridge regression using Rcpp and parallel programming. The package is demonstrated on an electricity demand dataset.
}
\author{
Euan Enticott, Georgina Mansell, and Conor Newton
}

