% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvn_omp}
\alias{rmvn_omp}
\title{Sample from a multivariate Gaussian}
\usage{
rmvn_omp(n, mu, sigma)
}
\arguments{
\item{n}{Integer number of samples to draw}

\item{mu}{Vector of means}

\item{sigma}{Covariance matrix}
}
\value{
Matrix of MVN samples (n rows)
}
\description{
Sample from a multivariate Gaussian
}
