% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{spectralClustering}
\alias{spectralClustering}
\title{Spectral clustering}
\usage{
spectralClustering(x, c = 1, k = 10)
}
\arguments{
\item{x}{First value}
}
\value{
vector of groups
}
\description{
Spectral clustering
}
