% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_ocv}
\alias{get_ocv}
\title{Calculate leave one out cross validation error (OCV)}
\usage{
get_ocv(X, y, lambda)
}
\arguments{
\item{X}{First value}

\item{y}{Second value}

\item{lambda}{}
}
\description{
Calculate leave one out cross validation error (OCV)
}
