% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optim_rr}
\alias{optim_rr}
\title{Find the optimal regularisation parameter through optimised leave one out cross validation}
\usage{
optim_rr(X, y, lams)
}
\arguments{
\item{X}{Data matrix}

\item{y}{Column matrix of responses}

\item{lams}{Vector of regularisation parameters to test}
}
\value{
Vector of OCVs
}
\description{
Find the optimal regularisation parameter through optimised leave one out cross validation
}
