% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{par_reg}
\alias{par_reg}
\title{Fit a ridge regression model to multiple groups in parallel}
\usage{
par_reg(X, y, lams, idx)
}
\arguments{
\item{X}{First value}

\item{y}{Second value}

\item{lams}{First value}

\item{idx}{Second value}
}
\value{
List
}
\description{
Fit a ridge regression model to multiple groups in parallel
}
