% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvn_omp}
\alias{rmvn_omp}
\title{Sample from a multivariate Gaussian}
\usage{
rmvn_omp(n, mu, sigma)
}
\arguments{
\item{n}{First value}

\item{mu}{Second value}
}
\value{
matrix
}
\description{
Sample from a multivariate Gaussian
}
